Fix DB utils: empty list guards, baseFilter consistency, cleanup

+++ src/main/scala/madrileno/utils/db/dsl/ForeignIdRepository.scala
@@ -17,12 +17,13 @@ trait ForeignIdRepository[A, Id] extends BaseRepository[A] {
       )(foreignId)
   }
 
-  def findByForeignIds(foreignIds: List[Id], lock: Lock = Lock.NoLock)(using session: Session[IO]): IO[List[A]] = {
-    session.execute(
-      sql"SELECT ${table.*} FROM ${table.n} WHERE $baseFilter AND ${table.foreignId.n} IN (${table.foreignId.c.list(foreignIds)}) ${lock.fragment}"
-        .query(table.c)
-    )(foreignIds)
-  }
+  def findByForeignIds(foreignIds: List[Id], lock: Lock = Lock.NoLock)(using session: Session[IO]): IO[List[A]] =
+    if (foreignIds.isEmpty) IO.pure(Nil)
+    else
+      session.execute(
+        sql"SELECT ${table.*} FROM ${table.n} WHERE $baseFilter AND ${table.foreignId.n} IN (${table.foreignId.c.list(foreignIds)}) ${lock.fragment}"
+          .query(table.c)
+      )(foreignIds)
 
   def countByForeignId(foreignId: Id)(using session: Session[IO]): IO[Long] = {
     session.unique(sql"SELECT COUNT(*) FROM ${table.n} WHERE $baseFilter AND ${table.foreignId.n} = ${table.foreignId.c}".query(int8))(foreignId)
@@ -39,9 +40,11 @@ trait ForeignIdRepository[A, Id] extends BaseRepository[A] {
       .void
 
   def deleteByForeignIds(foreignIds: List[Id])(using session: Session[IO]): IO[Unit] =
-    session
-      .execute(sql"DELETE FROM ${table.n} WHERE ${table.foreignId.n} IN (${table.foreignId.c.list(foreignIds)})".command)(foreignIds)
-      .void
+    if (foreignIds.isEmpty) IO.unit
+    else
+      session
+        .execute(sql"DELETE FROM ${table.n} WHERE ${table.foreignId.n} IN (${table.foreignId.c.list(foreignIds)})".command)(foreignIds)
+        .void
 
   override val table: Table[A] & ForeignIdTable[Id]
 }
